kafka_producer/requirements.txt
pandas
kafka-python==2.0.2

